@@ -8,7 +8,7 @@ import csv
 from datetime import datetime
 from random import randint
 
-from sefaria.system.exceptions import InputError
+from sefaria.system.exceptions import InputError, SheetNotFoundError
 
 if not hasattr(sys, '_doc_build'):
     from django.contrib.auth.models import User
@@ -56,7 +56,8 @@ class UserHistory(abst.AbstractMongoRecord):
         "num_times_read",     # int: legacy for migrating old recent views
         "sheet_title",        # str: for sheet history
         "sheet_owner",        # str: ditto
-        "sheet_id"            # int: ditto
+        "sheet_id",           # int: ditto
+        "delete_saved",       # bool: True if this item was saved and but then was deleted
     ]
 
     def __init__(self, attrs=None, load_existing=False, field_updates=None, update_last_place=False):
@@ -109,6 +110,11 @@ class UserHistory(abst.AbstractMongoRecord):
             if not self.secondary and not self.is_sheet and getattr(self, "language", None) != "hebrew" and r.is_empty("en"):
                 # logically, this would be on frontend, but easier here.
                 self.language = "hebrew"
+        except SheetNotFoundError:
+            self.context_refs   = [self.ref]
+            self.categories     = ["_unlisted"]
+            self.authors        = []
+            self.is_sheet       = True
         except InputError:   # Ref failed to resolve
             self.context_refs   = [self.ref]
             self.categories     = []
@@ -120,18 +126,22 @@ class UserHistory(abst.AbstractMongoRecord):
     def contents(self, **kwargs):
         d = super(UserHistory, self).contents(**kwargs)
         if kwargs.get("for_api", False):
-            try:
-                del d["uid"]
-            except KeyError:
-                pass
-            try:
-                del d["server_time_stamp"]
-            except KeyError:
-                pass
-            try:
-                d["datetime"] = str(d["datetime"])
-            except KeyError:
-                pass
+            keys = {
+                'ref': u'',
+                'he_ref': u'',
+                'book': u'',
+                'versions': {},
+                'time_stamp': 0,
+                'saved': False,
+                'delete_saved': False,
+                'is_sheet': False,
+                'sheet_id': -1,
+                'sheet_owner': '',
+                'sheet_title': '',
+            }
+            d = {
+                key: d.get(key, default) for key, default in keys.items()
+            }
         return d
 
     def _sanitize(self):
@@ -153,7 +163,8 @@ class UserHistory(abst.AbstractMongoRecord):
         saved = True if action == "add_saved" else (False if action == "delete_saved" else hist.get("saved", False))
         uh = UserHistory(hist, load_existing=(action is not None), update_last_place=(action is None), field_updates={
             "saved": saved,
-            "server_time_stamp": hist["server_time_stamp"]
+            "server_time_stamp": hist["server_time_stamp"],
+            "delete_saved": action == "delete_saved"
         })
         uh.save()
         return uh
@@ -201,6 +212,9 @@ class UserHistory(abst.AbstractMongoRecord):
 
 class UserHistorySet(abst.AbstractMongoSet):
     recordClass = UserHistory
+
+    def hits(self):
+        return reduce(lambda agg,o: agg + getattr(o, "num_times_read", 1), self, 0)
 
 
 class UserProfile(object):
@@ -338,24 +352,11 @@ class UserProfile(object):
             self.save()
         return profile
 
-    def update_attr_time_stamps(self, obj):
-        if "settings" in obj:
-            settings_changed = False
-            for k, v in obj["settings"].items():
-                if k not in self.settings:
-                    settings_changed = True
-                elif v != self.settings[k]:
-                    settings_changed = True
-            if settings_changed:
-                obj["attr_time_stamps"] = obj.get("attr_time_stamps", {})
-                obj["attr_time_stamps"]["settings"] = epoch_time()
-
     def update(self, obj):
         """
         Update this object with the fields in dictionry 'obj'
         """
         self._set_flags_on_update(obj)
-        self.update_attr_time_stamps(obj)
         self.__dict__.update(obj)
 
         return self
@@ -546,7 +547,7 @@ class UserProfile(object):
     def to_api_dict(self, basic=False):
         """
         Return a json serializble dictionary this profile which includes fields used in profile API methods
-        If basic is True, only return enough data to display a profile listing 
+        If basic is True, only return enough data to display a profile listing
         """
         if basic:
             return {
@@ -662,111 +663,6 @@ def public_user_data(uid):
     public_user_data_cache[uid] = data
     return data
 
-# Properly, this doesn't belong here, but it's being developed alongside user_stats_data
-# todo: relocate the stats methods?
-def site_stats_data():
-    from sefaria.model.category import TOP_CATEGORIES
-    from sefaria.model.trend import Trend, TrendSet, read_in_category_key, reverse_read_in_category_key
-
-    d = {"categoriesRead": {reverse_read_in_category_key(t.name): t.value
-                            for t in TrendSet({"scope": "site", "period": "alltime",
-                                "name": {"$in": map(read_in_category_key, TOP_CATEGORIES)}
-                            })}}
-    return d
-
-
-def user_stats_data(uid, start=None, end=None):
-    """
-
-    :param uid: int or something cast-able to int
-    :param start: datetime
-    :param end: datetime
-    :return:
-    """
-    from sefaria.model.category import TOP_CATEGORIES
-    from sefaria.model.trend import Trend, TrendSet, read_in_category_key, reverse_read_in_category_key
-    from sefaria.sheets import user_sheets, sheet_list
-
-    uid = int(uid)
-
-    # todo: needs more thought.  UserHistory.timeclause handles the Nones, but later usages in this method aren't so graceful.
-    timeclause = UserHistory.timeclause(start, end)
-    end = end or datetime.now()
-    start = start or datetime(2017, 12, 1)  # start of Sefaria epoch
-
-    # All of user's sheets
-    usheets = user_sheets(uid)["sheets"]
-    usheet_ids = [s["id"] for s in usheets]
-
-    # Sheet views in this period
-    match_clause = {
-            "is_sheet": True,
-            "sheet_id": {"$in": usheet_ids},
-            }
-    match_clause.update(timeclause)
-    usheet_views = db.user_history.aggregate([
-        {"$match": match_clause},
-        {"$group": {
-            "_id": "$sheet_id",
-            "cnt": {"$sum": 1}}},
-    ])
-
-    most_popular_sheet_ids = [s["_id"] for s in sorted(usheet_views, key=lambda o: o["cnt"], reverse=True)[:3]]
-    most_popular_sheets = [s for s in usheets if s["id"] in most_popular_sheet_ids]
-    sheets_this_period = [s for s in usheets if start <= datetime.strptime(s["created"], "%Y-%m-%dT%H:%M:%S.%f") <= end]
-
-    # Refs I viewed
-    match_clause = {
-            "uid": uid,
-            "secondary": False,
-            "is_sheet": False
-            }
-    match_clause.update(timeclause)
-    refs_viewed = db.user_history.aggregate([
-        {"$match": match_clause},
-        {"$group": {
-            "_id": "$ref",
-            "cnt": {"$sum": 1}}},
-    ])
-    most_viewed_refs = [s["_id"] for s in sorted(refs_viewed, key=lambda o: o["cnt"], reverse=True) if s["cnt"] > 1 and "Genesis 1" not in s["_id"]][:10]
-
-
-    # Sheets I viewed
-    match_clause = {
-            "uid": uid,
-            "secondary": False,
-            "is_sheet": True
-            }
-    match_clause.update(timeclause)
-    sheets_viewed = db.user_history.aggregate([
-        {"$match": match_clause},
-        {"$group": {
-            "_id": "$sheet_id",
-            "cnt": {"$sum": 1}}},
-    ])
-    most_viewed_sheets_ids = [s["_id"] for s in sorted(sheets_viewed, key=lambda o: o["cnt"], reverse=True) if s["cnt"] > 1 and s["_id"] not in usheet_ids][:10]
-    most_viewed_sheets = sheet_list({"id": {"$in":most_viewed_sheets_ids}})
-
-    # Construct returned data
-    d = public_user_data(uid)
-
-    sheetsReadQuery = {"is_sheet": True, "secondary": False, "uid": uid}
-    sheetsReadQuery.update(timeclause)
-    d["sheetsRead"] = UserHistorySet(sheetsReadQuery).count()
-
-    textsReadQuery = {"is_sheet": False, "secondary": False, "uid": uid}
-    textsReadQuery.update(timeclause)
-    d["textsRead"] = UserHistorySet(textsReadQuery).count()
-
-    d["categoriesRead"] = {reverse_read_in_category_key(t.name): t.value for t in TrendSet({"uid":uid, "name": {"$in": map(read_in_category_key, TOP_CATEGORIES)}})}
-    d["totalSheets"] = len(usheets)
-    d["publicSheets"] = len([s for s in usheets if s["status"] == "public"])
-    d["popularSheets"] = most_popular_sheets
-    d["sheetsThisPeriod"] = len(sheets_this_period)
-    d["mostViewedRefs"] = most_viewed_refs
-    d["mostViewedSheets"] = most_viewed_sheets
-    return d
-
 
 def user_name(uid):
     """Returns a string of a user's full name"""
@@ -848,4 +744,3 @@ def process_index_title_change_in_user_history(indx, **kwargs):
             o.save()
         except InputError:
             logger.warning(u"Failed to convert user history from: {} to {}".format(kwargs['old'], kwargs['new']))
-
