@@ -1,7 +1,6 @@
 # Django settings for sefaria project.
 
 import os.path
-
 from django.utils.translation import ugettext_lazy as _
 
 relative_to_abs_path = lambda *x: os.path.join(os.path.dirname(
@@ -149,6 +148,7 @@ INSTALLED_APPS = (
     'anymail',
     'webpack_loader',
     'django_user_agents',
+    'rest_framework',
     #'easy_timezones'
     # Uncomment the next line to enable admin documentation:
     # 'django.contrib.admindocs',
@@ -164,6 +164,12 @@ AUTHENTICATION_BACKENDS = (
     'emailusernames.backends.EmailAuthBackend',
 )
 
+REST_FRAMEWORK = {
+    'DEFAULT_AUTHENTICATION_CLASSES': (
+        'rest_framework_simplejwt.authentication.JWTAuthentication',
+        'rest_framework.authentication.SessionAuthentication',
+    )
+}
 
 LOCALE_PATHS = (
     relative_to_abs_path('../locale'),
