@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-
+from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
 from django.conf.urls import include, url
 from django.conf.urls import handler404, handler500
 from django.contrib import admin
@@ -51,7 +51,7 @@ urlpatterns = [
     url(r'^modtools/?$', reader_views.modtools),
     url(r'^new-home/?$', reader_views.new_home),
     url(r'^story_editor/?$', reader_views.story_editor),
-    url(r'^user_stats/?$', reader_views.user_stats),
+    url(r'^torahtracker/?$', reader_views.user_stats),
 
 ]
 
@@ -304,6 +304,9 @@ urlpatterns += [
     url(r'^password/reset/confirm/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$', django_auth_views.PasswordResetConfirmView.as_view(), name='password_reset_confirm'),
     url(r'^password/reset/complete/$', django_auth_views.PasswordResetCompleteView.as_view(), name='password_reset_complete'),
     url(r'^password/reset/done/$', django_auth_views.PasswordResetDoneView.as_view(), name='password_reset_done'),
+    url(r'^api/register/$', sefaria_views.register_api),
+    url(r'^api/login/$', TokenObtainPairView.as_view(), name='token_obtain_pair'),
+    url(r'^api/login/refresh/$', TokenRefreshView.as_view(), name='token_refresh'),
 ]
 
 # Compare Page
